+++ tests/presets.py
@@ -69,3 +69,16 @@ ALLOWED_SCHEMES_DEFAULT = {
 ALLOWED_SCHEMES_HTTP = {
     "ALLOWED_SCHEMES": ["https", "http"],
 }
+
+DCR_SETTINGS = {
+    "SCOPES": {
+        "read": "Read scope",
+        "write": "Write scope",
+    },
+    "READ_SCOPE": "read",
+    "WRITE_SCOPE": "write",
+    "DCR_ENABLED": True,
+    "DCR_REGISTRATION_PERMISSION_CLASSES": ("oauth2_provider.dcr.IsAuthenticatedDCRPermission",),
+    "DCR_REGISTRATION_SCOPE": "oauth2_provider:registration",
+    "ALLOWED_REDIRECT_URI_SCHEMES": ["https", "http"],
+}